backend/requirements.txt
pydantic
pydantic-settings
httpx
selenium-wire
rich
fastapi
uvicorn
SQLAlchemy
pony
pymongo
beanie
motor
python-decouple
langchain
openai
tiktoken
faiss-cpu
redis
asyncio
spacy
pytest
pytest-asyncio
asgi-lifespan
scikit-learn
fake-useragent
streamlit
watchdog
plotly
